@@ -6,15 +6,13 @@
     "@testing-library/jest-dom": "^5.11.4",
     "@testing-library/react": "^11.1.0",
     "@testing-library/user-event": "^12.1.10",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-hook-form": "^7.27.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-redux": "^7.2.4",
     "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
+    "react-scripts": "^5.0.1",
     "redux": "^4.1.1",
     "redux-thunk": "^2.3.0",
-    "searchkit": "2.3.0",
     "web-vitals": "^1.0.1"
   },
   "scripts": {
@@ -42,6 +40,6 @@
     ]
   },
   "devDependencies": {
-    "node-sass": "4"
+    "node-sass": "^9.0.0"
   }
 }
