@@ -1,7 +1,4 @@
-/*
-    render props 作用类似HOC
-    组件逻辑复用
-*/ 
+
 
 
 import React,{useState} from 'react'
@@ -11,8 +8,8 @@ function Toggle(props) {
     const {render} = props
     return (
         <div>
-            {visible &&
-                render({visible})
+            {
+                render(visible)
             }
             <button onClick={e=>{
                 setVisible(!visible)
