@@ -1,41 +1,54 @@
 
 import React, { useState,useEffect, useReducer } from 'react'
 
+
+const initState = {
+    data: null,
+    loading: false,
+    error: null
+}
+
+/*
+    将所有状态集中管理，对状态的修改逻辑也集中在reducer管理，
+    可避免遗漏，方便维护
+
+    当组件状态比较复杂时或者是深层嵌套，建议采用useReducer替代useState.
+*/
 const reducer = (state,action)=>{
     switch(action.type){
-        case 'tick':
-            return {...state, count:state.count+state.step}
-        case 'changeStep':
-            return {...state, step:state.step+action.step}
+        case 'start':
+            return {...state, loading:true}
+        case 'success':
+            return {...state, data:action.payload, loading:false}
         default:
             return state;
     }
 }
 
 function TuseReducer() {
-    // const [count, setCount] = useState(0)
-    // const [step, setStep] = useState(1)
 
-    //当step改变时，会开启一个新的计时器
-    // useEffect(()=>{
-    //     setInterval(()=>setCount(count+step),1000)
-    // },[step])
-
-    const [state, dispatch] = useReducer(reducer, {count:0, step:1})
-    const {count, step} = state
-    
-    useEffect(()=>{
-        console.log('effect',count, step)
-        setInterval(()=>dispatch({type:'tick'}),1000)
-    },[dispatch])
+    // useReducer相当于一个小型的redux, 采用dispatch来更新状态
+    const [state, dispatch] = useReducer(reducer, initState)
+ 
     return (
         <div>
-            <p>count: {state.count}</p>
-            <p>step: {state.step}</p>
+            {state.loading && 
+                <p>loading ...</p>
+            }
+            {!state.loading && state.data && 
+                <p>{state.data}</p>
+            }
             <button onClick={()=>{
-                // setStep(step+1)
-                dispatch({type:'changeStep',step:1})
-            }}>step + 1</button>
+                dispatch({
+                    type: 'start',
+                })
+                setTimeout(()=>{
+                    dispatch({
+                        type: 'success',
+                        payload: 'Yes, you got it'
+                    })
+                },3000)
+            }}>获取数据</button>
         </div>
     )
 }
